fix compile error in verbose mode

+++ esphome/components/modbus_controller/output/modbus_output.cpp
@@ -54,7 +54,7 @@ void ModbusFloatOutput::dump_config() {
   LOG_FLOAT_OUTPUT(this);
   ESP_LOGCONFIG(TAG, "  Device start address=0x%X", this->start_address);
   ESP_LOGCONFIG(TAG, "  Register count=0x%d", this->register_count);
-  ESP_LOGCONFIG(TAG, "  Value type=%hhu", this->sensor_value_type);
+  ESP_LOGCONFIG(TAG, "  Value type=%hhu", static_cast<int>(this->sensor_value_type));
 }
 
 // ModbusBinaryOutput
@@ -62,6 +62,7 @@ void ModbusBinaryOutput::write_state(bool state) {
   // This will be called every time the user requests a state change.
   ModbusCommandItem cmd;
   std::vector<uint8_t> data;
+
   // Is there are lambda configured?
   if (this->write_transform_func_.has_value()) {
     // data is passed by reference
@@ -103,7 +104,7 @@ void ModbusBinaryOutput::dump_config() {
   LOG_BINARY_OUTPUT(this);
   ESP_LOGCONFIG(TAG, "  Device start address=0x%X", this->start_address);
   ESP_LOGCONFIG(TAG, "  Register count=0x%d", this->register_count);
-  ESP_LOGCONFIG(TAG, "  Value type=%hhu", this->sensor_value_type);
+  ESP_LOGCONFIG(TAG, "  Value type=%hhu", static_cast<int>(this->sensor_value_type));
 }
 
 }  // namespace modbus_controller
